Rename db.py to database_connection.py
